+++ R/vosonSML-package.R
@@ -35,8 +35,11 @@
 #' @importFrom RedditExtractoR reddit_content user_network
 #' @importFrom magrittr '%>%' '%<>%'
 #' @importFrom dplyr rename group_by summarise ungroup left_join select mutate filter coalesce row_number
-#' distinct anti_join mutate_all mutate_at ends_with vars funs bind_rows arrange
+#' distinct anti_join mutate_all mutate_at ends_with vars funs bind_rows arrange case_when if_else
+#' contains n rowwise
 #' @importFrom rlang '.data'
 #' @importFrom textutils HTMLdecode
 #' @importFrom tictoc tic toc
+#' @importFrom tibble as_tibble enframe
+#' @importFrom stats na.omit
 NULL